Replaced user methods to user controller
